Rename county audit wizard submodule.
